implement Into<Vec<Message>> for QueryError and QueryEvent (#182)
