+++ src/protocol/src/lib.rs
@@ -22,7 +22,7 @@ use bytes::BytesMut;
 use futures_util::io::{AsyncReadExt, AsyncWriteExt};
 use std::io;
 
-use crate::results::{QueryEvent, QueryResult};
+use crate::results::QueryResult;
 pub use listener::{QueryListener, ServerListener};
 
 /// Module contains functionality to listen to incoming client connections and
@@ -133,7 +133,8 @@ impl<RW: AsyncReadExt + AsyncWriteExt + Unpin> Connection<RW> {
     /// Sends response messages to client. Most of the time it is a single
     /// message, select result one of the exceptional situation
     pub async fn send(&mut self, query_result: QueryResult) -> io::Result<()> {
-        for message in QueryResultMapper::map(query_result) {
+        let messages: Vec<Message> = query_result.map_or_else(|event| event.into(), |err| err.into());
+        for message in messages {
             log::debug!("{:?}", message);
             self.socket.write_all(message.as_vec().as_slice()).await?;
         }
@@ -181,323 +182,9 @@ pub enum SslMode {
     Disable,
 }
 
-struct QueryResultMapper;
-
-impl QueryResultMapper {
-    fn map(resp: QueryResult) -> Vec<Message> {
-        match resp {
-            Ok(QueryEvent::SchemaCreated) => vec![Message::CommandComplete("CREATE SCHEMA".to_owned())],
-            Ok(QueryEvent::SchemaDropped) => vec![Message::CommandComplete("DROP SCHEMA".to_owned())],
-            Ok(QueryEvent::TableCreated) => vec![Message::CommandComplete("CREATE TABLE".to_owned())],
-            Ok(QueryEvent::TableDropped) => vec![Message::CommandComplete("DROP TABLE".to_owned())],
-            Ok(QueryEvent::VariableSet) => vec![Message::CommandComplete("SET".to_owned())],
-            Ok(QueryEvent::TransactionStarted) => vec![Message::CommandComplete("BEGIN".to_owned())],
-            Ok(QueryEvent::RecordsInserted(records)) => vec![Message::CommandComplete(format!("INSERT 0 {}", records))],
-            Ok(QueryEvent::RecordsSelected(projection)) => {
-                let definition = projection.0;
-                let description: Vec<ColumnMetadata> = definition
-                    .into_iter()
-                    .map(|(name, sql_type)| ColumnMetadata::new(name, sql_type.pg_oid(), sql_type.pg_len()))
-                    .collect();
-                let records = projection.1;
-                let len = records.len();
-                let mut messages = vec![Message::RowDescription(description)];
-                for record in records {
-                    messages.push(Message::DataRow(record));
-                }
-                messages.push(Message::CommandComplete(format!("SELECT {}", len)));
-                messages
-            }
-            Ok(QueryEvent::RecordsUpdated(records)) => vec![Message::CommandComplete(format!("UPDATE {}", records))],
-            Ok(QueryEvent::RecordsDeleted(records)) => vec![Message::CommandComplete(format!("DELETE {}", records))],
-            Err(query_error) => query_error.into_messages(),
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[cfg(test)]
-    mod mapper {
-        use super::*;
-        use crate::{results::QueryErrorBuilder, sql_types::PostgreSqlType};
-
-        #[test]
-        fn create_schema() {
-            assert_eq!(
-                QueryResultMapper::map(Ok(QueryEvent::SchemaCreated)),
-                vec![Message::CommandComplete("CREATE SCHEMA".to_owned())]
-            )
-        }
-
-        #[test]
-        fn drop_schema() {
-            assert_eq!(
-                QueryResultMapper::map(Ok(QueryEvent::SchemaDropped)),
-                vec![Message::CommandComplete("DROP SCHEMA".to_owned())]
-            )
-        }
-
-        #[test]
-        fn create_table() {
-            assert_eq!(
-                QueryResultMapper::map(Ok(QueryEvent::TableCreated)),
-                vec![Message::CommandComplete("CREATE TABLE".to_owned())]
-            );
-        }
-
-        #[test]
-        fn drop_table() {
-            assert_eq!(
-                QueryResultMapper::map(Ok(QueryEvent::TableDropped)),
-                vec![Message::CommandComplete("DROP TABLE".to_owned())]
-            );
-        }
-
-        #[test]
-        fn insert_record() {
-            let records_number = 3;
-            assert_eq!(
-                QueryResultMapper::map(Ok(QueryEvent::RecordsInserted(records_number))),
-                vec![Message::CommandComplete(format!("INSERT 0 {}", records_number))]
-            )
-        }
-
-        #[test]
-        fn select_records() {
-            let projection = (
-                vec![
-                    ("column_name_1".to_owned(), PostgreSqlType::SmallInt),
-                    ("column_name_2".to_owned(), PostgreSqlType::SmallInt),
-                ],
-                vec![
-                    vec!["1".to_owned(), "2".to_owned()],
-                    vec!["3".to_owned(), "4".to_owned()],
-                ],
-            );
-            assert_eq!(
-                QueryResultMapper::map(Ok(QueryEvent::RecordsSelected(projection))),
-                vec![
-                    Message::RowDescription(vec![
-                        ColumnMetadata::new("column_name_1".to_owned(), 21, 2),
-                        ColumnMetadata::new("column_name_2".to_owned(), 21, 2)
-                    ]),
-                    Message::DataRow(vec!["1".to_owned(), "2".to_owned()]),
-                    Message::DataRow(vec!["3".to_owned(), "4".to_owned()]),
-                    Message::CommandComplete("SELECT 2".to_owned())
-                ]
-            );
-        }
-
-        #[test]
-        fn update_records() {
-            let records_number = 3;
-            assert_eq!(
-                QueryResultMapper::map(Ok(QueryEvent::RecordsUpdated(records_number))),
-                vec![Message::CommandComplete(format!("UPDATE {}", records_number))]
-            );
-        }
-
-        #[test]
-        fn delete_records() {
-            let records_number = 3;
-            assert_eq!(
-                QueryResultMapper::map(Ok(QueryEvent::RecordsDeleted(records_number))),
-                vec![Message::CommandComplete(format!("DELETE {}", records_number))]
-            )
-        }
-
-        #[test]
-        fn schema_already_exists() {
-            let schema_name = "some_table_name".to_owned();
-            assert_eq!(
-                QueryResultMapper::map(Err(QueryErrorBuilder::new()
-                    .schema_already_exists(schema_name.clone())
-                    .build())),
-                vec![Message::ErrorResponse(
-                    Some("ERROR"),
-                    Some("42P06"),
-                    Some(format!("schema \"{}\" already exists", schema_name)),
-                )]
-            )
-        }
-
-        #[test]
-        fn schema_does_not_exists() {
-            let schema_name = "some_table_name".to_owned();
-            assert_eq!(
-                QueryResultMapper::map(Err(QueryErrorBuilder::new()
-                    .schema_does_not_exist(schema_name.clone())
-                    .build())),
-                vec![Message::ErrorResponse(
-                    Some("ERROR"),
-                    Some("3F000"),
-                    Some(format!("schema \"{}\" does not exist", schema_name)),
-                )]
-            )
-        }
-
-        #[test]
-        fn table_already_exists() {
-            let table_name = "some_table_name".to_owned();
-            assert_eq!(
-                QueryResultMapper::map(Err(QueryErrorBuilder::new()
-                    .table_already_exists(table_name.clone())
-                    .build())),
-                vec![Message::ErrorResponse(
-                    Some("ERROR"),
-                    Some("42P07"),
-                    Some(format!("table \"{}\" already exists", table_name)),
-                )]
-            )
-        }
-
-        #[test]
-        fn table_does_not_exists() {
-            let table_name = "some_table_name".to_owned();
-            assert_eq!(
-                QueryResultMapper::map(Err(QueryErrorBuilder::new()
-                    .table_does_not_exist(table_name.clone())
-                    .build())),
-                vec![Message::ErrorResponse(
-                    Some("ERROR"),
-                    Some("42P01"),
-                    Some(format!("table \"{}\" does not exist", table_name)),
-                )]
-            )
-        }
-
-        #[test]
-        fn one_column_does_not_exists() {
-            assert_eq!(
-                QueryResultMapper::map(Err(QueryErrorBuilder::new()
-                    .column_does_not_exist(vec!["column_not_in_table".to_owned()])
-                    .build())),
-                vec![Message::ErrorResponse(
-                    Some("ERROR"),
-                    Some("42703"),
-                    Some("column column_not_in_table does not exist".to_owned()),
-                )]
-            )
-        }
-
-        #[test]
-        fn multiple_columns_does_not_exists() {
-            assert_eq!(
-                QueryResultMapper::map(Err(QueryErrorBuilder::new()
-                    .column_does_not_exist(vec![
-                        "column_not_in_table1".to_owned(),
-                        "column_not_in_table2".to_owned()
-                    ])
-                    .build())),
-                vec![Message::ErrorResponse(
-                    Some("ERROR"),
-                    Some("42703"),
-                    Some("columns column_not_in_table1, column_not_in_table2 do not exist".to_owned()),
-                )]
-            )
-        }
-
-        #[test]
-        fn feature_not_supported() {
-            let raw_sql_query = "some SQL query".to_owned();
-            assert_eq!(
-                QueryResultMapper::map(Err(QueryErrorBuilder::new()
-                    .feature_not_supported(raw_sql_query.clone())
-                    .build())),
-                vec![Message::ErrorResponse(
-                    Some("ERROR"),
-                    Some("0A000"),
-                    Some(format!("Currently, Query '{}' can't be executed", raw_sql_query)),
-                )]
-            )
-        }
-
-        #[test]
-        fn too_many_insert_expressions() {
-            assert_eq!(
-                QueryResultMapper::map(Err(QueryErrorBuilder::new().too_many_insert_expressions().build())),
-                vec![Message::ErrorResponse(
-                    Some("ERROR"),
-                    Some("42601"),
-                    Some("INSERT has more expressions than target columns".to_owned()),
-                )]
-            )
-        }
-
-        #[test]
-        fn out_of_range_constraint_violation() {
-            let mut builder = QueryErrorBuilder::new();
-            builder.out_of_range(PostgreSqlType::SmallInt, "col1".to_string(), 1);
-
-            assert_eq!(
-                QueryResultMapper::map(Err(builder.build())),
-                vec![Message::ErrorResponse(
-                    Some("ERROR"),
-                    Some("22003"),
-                    Some("smallint is out of range for column 'col1' at row 1".to_owned())
-                )]
-            )
-        }
-
-        #[test]
-        fn type_mismatch_constraint_violation() {
-            let mut builder = QueryErrorBuilder::new();
-            builder.type_mismatch("abc", PostgreSqlType::SmallInt, "col1".to_string(), 1);
-            assert_eq!(
-                QueryResultMapper::map(Err(builder.build())),
-                vec![Message::ErrorResponse(
-                    Some("ERROR"),
-                    Some("2200G"),
-                    Some("invalid input syntax for type smallint for column 'col1' at row 1: \"abc\"".to_owned())
-                )]
-            )
-        }
-
-        #[test]
-        fn string_length_mismatch_constraint_violation() {
-            let mut builder = QueryErrorBuilder::new();
-            builder.string_length_mismatch(PostgreSqlType::Char, 5, "col1".to_string(), 1);
-            assert_eq!(
-                QueryResultMapper::map(Err(builder.build())),
-                vec![Message::ErrorResponse(
-                    Some("ERROR"),
-                    Some("22026"),
-                    Some("value too long for type character(5) for column 'col1' at row 1".to_owned())
-                )]
-            )
-        }
-
-        #[test]
-        fn undefined_function() {
-            assert_eq!(
-                QueryResultMapper::map(Err(QueryErrorBuilder::new()
-                    .undefined_function("||".to_owned(), "NUMBER".to_owned(), "NUMBER".to_owned())
-                    .build())),
-                vec![Message::ErrorResponse(
-                    Some("ERROR"),
-                    Some("42883"),
-                    Some("operator does not exist: (NUMBER || NUMBER)".to_owned())
-                )]
-            )
-        }
-
-        #[test]
-        fn syntax_error() {
-            assert_eq!(
-                QueryResultMapper::map(Err(QueryErrorBuilder::new()
-                    .syntax_error("expression".to_owned())
-                    .build())),
-                vec![Message::ErrorResponse(
-                    Some("ERROR"),
-                    Some("42601"),
-                    Some("syntax error in expression".to_owned())
-                )]
-            )
-        }
-    }
 
     #[cfg(test)]
     mod connection {
